Developed solution for on screen character map navigation in 3 languages
This is the solution from Daryl Shockey. It includes a readme document in Word format called "CharacterMappingReadMe.docx". Please take a moment to read this document as it describes the solutions and methodology. There are also sample inputs, configuration files, and result sets in the "SampleKeyboardData" folder. All of the final executable products, in release form, can be found in the "Binaries" folder. All of the source content is in each respective project folder.